perf: 优化分页获取商品列表和分页查询商品接口
用mysql的explain优化查询商品大致数量

+++ mall-admin/src/main/java/online/ahayujie/mall/admin/pms/controller/ProductController.java
@@ -153,8 +153,8 @@ public class ProductController {
     @ApiOperation(value = "分页查询商品", notes = "根据商品名称，货号，分类，品牌，上架状态，新品状态，推荐状态，审核状态，预告状态查询")
     @PostMapping("/query/{pageNum}/{pageSize}")
     public Result<CommonPage<Product>> query(@RequestBody QueryProductParam param,
-                                             @PathVariable Integer pageNum,
-                                             @PathVariable Integer pageSize) {
+                                             @PathVariable Long pageNum,
+                                             @PathVariable Long pageSize) {
         return Result.data(productService.queryProduct(param, pageNum, pageSize));
     }
 }
